Attempting to clarify the whole read/write/close situation in WvStreams/Win32. Removed source file from project file to make ExchangeIt/Outlook auto-builder work again (maybe).
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@3910 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/wvtimeutils.h
@@ -20,9 +20,9 @@ public:
     WvTime()
         { } // WARNING: leaves members uninitialized, like timeval would do!
     WvTime(long long t)
-        { tv_sec = t/1000000L; tv_usec = t%1000000L; }
+        { tv_sec = long(t/1000000L); tv_usec = long(t%1000000L); }
     WvTime(time_t sec, time_t usec)
-        { tv_sec = sec; tv_usec = usec; }
+        { tv_sec = long(sec); tv_usec = long(usec); }
     WvTime(const struct timeval &tv)
         { tv_sec = tv.tv_sec; tv_usec = tv.tv_usec; }
     WvTime(const WvTime &tv)
